+++ xenian_bot/commands/calculator.py
@@ -3,7 +3,8 @@ import re
 from math import *
 
 from telegram import Bot, ParseMode, Update
-from telegram.ext import Filters, MessageHandler
+from telegram.error import BadRequest
+from telegram.ext import Filters, MessageHandler, run_async
 
 from . import BaseCommand
 
@@ -29,35 +30,76 @@ class Calculator(BaseCommand):
         self.commands = [
             {
                 'title': 'Calculator',
-                'description': 'Solve equations you send me, to get a full list of supported math functions use /maths',
+                'description': 'Solve equations you send me, to get a full list of supported math functions use /maths '
+                               '(PRIVATE CHAT ONLY)',
                 'command': self.calcualate,
                 'handler': MessageHandler,
-                'options': {'filters': Filters.text},
+                'options': {'filters': Filters.text & ~ Filters.group},
             },
             {
                 'title': 'Math Functions',
                 'description': 'Show all available math functions',
                 'command': self.maths,
             },
+            {
+                'title': 'Calculate',
+                'description': 'Solve an equation you send me, all math functions can be seen with /maths',
+                'args': 'EQUATION',
+                'command_name': 'calc',
+                'options': {'pass_args': True},
+                'command': self.calcualate_command,
+            },
         ]
 
         super(Calculator, self).__init__()
 
-    def calcualate(self, bot: Bot, update: Update):
+    @run_async
+    def calcualate(self, bot: Bot, update: Update, equation: str = None):
         """Calculate
 
         Args:
             bot (:obj:`telegram.bot.Bot`): Telegram Api Bot Object.
             update (:obj:`telegram.update.Update`): Telegram Api Update Object
+            equation (:obj:`str`): Give an equation to solve and send the user or try to use the sent text
         """
-        equation = update.message.text
+        equation = equation or update.message.text
         equation = re.sub('["\']', '', equation)
+        equation = re.sub('(true|false)', '', equation, flags=re.IGNORECASE)
+        message = ''
         try:
             result = eval(equation, {"__builtins__": None}, self.safe_dict)
-            update.message.reply_text('`{} = {}`'.format(equation, result), parse_mode=ParseMode.MARKDOWN)
+            reply_template = '{message}\n`{equation} = {result}`'
+            try:
+                result = float(result)
+            except OverflowError:
+                message = 'Result could not be shortened, shown in long format:'
+            try:
+                reply = reply_template.format(
+                    message=message,
+                    equation=equation,
+                    result=result
+                )
+                update.message.reply_text(reply, parse_mode=ParseMode.MARKDOWN)
+            except BadRequest:
+                update.message.reply_text('Result was too long (max. 4096 characters) for Telegram.')
         except (SyntaxError, TypeError) as e:
             pass
 
+    def calcualate_command(self, bot: Bot, update: Update, args: list = None):
+        """Calculate
+
+        Args:
+            bot (:obj:`telegram.bot.Bot`): Telegram Api Bot Object.
+            update (:obj:`telegram.update.Update`): Telegram Api Update Object
+            args (:obj:`List`): User shall pass a equation
+        """
+        equation = ' '.join(args).strip()
+        if equation:
+            self.calcualate(bot, update, equation)
+        else:
+            update.message.reply_text('You have to give me an equation')
+
+    @run_async
     def maths(self, bot: Bot, update: Update):
         """Show available Maths Functions