+++ CHANGELOG.md
@@ -13,3 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add command `/rules_define YOUR_RULES` to define new rules in a group
 - Add command `/rules_remvoe` to remove the groups rules
 - Specify a time until user can return from kick with `/kick [TIME]`
+- Add `/calc EQUATION` command to calculate equations inside groups
+
+### Changed
+- Run math function asynchronous
+- Disable directly solving equations without command sent to groups
+- Fix not shortening solutions form the calculator
+- Fix message too long for Telegram, for too long solutions from the calculator
+- Remove all `True` and `False` before trying to calculate so a message with just "true" doesn't get returned